add CRUD for social Assistance recipient resource
- implement getAll, getAllPaginate, getByID, create, update, delete methods in repository & controller
- add request validations for store and update
- add API resource for response formatting

+++ database/migrations/2025_05_29_153209_create_social_assistance_recipients_table.php
@@ -29,7 +29,7 @@ return new class extends Migration
             $table->decimal('amount', 10, 2);
             $table->longText('reason');
             $table->enum('bank', ['bri', 'bni', 'mandiri', 'bca', 'other']);
-            $table->integer('bank_account_number');
+            $table->bigInteger('bank_account_number');
             $table->string('proof');
             $table->enum('status', ['pending', 'approved', 'rejected'])->default('pending');
